requirements.txt
# Core Dependencies
python-kraken-sdk==2.3.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0

# Google Cloud Platform
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.38.0
google-cloud-storage>=3.4.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.14.1

# Data Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# API & Data Fetching
requests>=2.31.0
websocket-client>=1.6.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
flake8>=6.1.0
black>=23.9.0

# Jupyter (for experimentation)
jupyter>=1.0.0
ipykernel>=6.25.0

